Optimization of VirtualHereFree Process Management
The vhfwatchdog script has been improved to check not only the 'streaming_client' process but also the 'moonlight' process. If one of these two processes is running and the 'vhusbdarm' process is not, then 'vhusbdarm' is started. Conversely, if neither the 'streaming_client' nor the 'moonlight' process is running while 'vhusbdarm' is active, then 'vhusbdarm' is stopped.